+++ NEWS.md
@@ -1,11 +1,13 @@
 # happign 0.1.4
 
-* Fix resolution for `get_wms_raster()`. Depending on shape and resolution, multiple tile are downloaded and combine to get the right resolution
+* Fix resolution for `get_wms_raster()`. Depending on shape and resolution, multiple tile are downloaded and combine to get the right resolution. Also adding vignette [Resolution for raster](https://paul-carteron.github.io/happign/articles/web_only/resolution_for_raster.html) for further explanation
 * New start up message based on RSS flux of IGN website to warn user if there issues (slowdown, shutdown) or news resources
 *`get_wms_raster()` now fix S2 geometry problems
 * adding `method` and `mode` argument of `download.file()` to have more freedom on the type of download with `get_wms_raster()`
 * Completion of the `happign_for_forester` vignette
 * adding first `get_apicarto_*` vectorized function for cadastre
+* adding `shp_to_geojson()` function to avoid `geojsonsf` package dependency
+
 
 # happign 0.1.3